Add to cart functionality
Created the logic that adds items to the cart useState. To accomplish this I needed to implement a number of functions.

handleSubmit: This function takes the form inputs and turns them into a all encomopassing object. The quantity property is turned into a number with the buil in js function Number(). The new object is then used as an arguement for the scanItem function.

scanItem: This function checks to see if the object quantity property is greater than 1. If it is a variable is created and is equal to the result of the duplicateItem function. Once the array is returned, it is used as an arguement to the addToCart function.

duplicateItem: This function creates an empty array on every run. The function takes the object and runs a for loop based on the objects quantity property. Within the for loop it will duplicate the object but set its' quantity property to 1. And then push the item to the array. Upon the for loop comppletion, the array is returned.

addToCart: This function checks to see if the arguement provided is an array. If it is array is run through a .map function and updates the useState with all of the objects. If the arguement is not an array, it just updates the useState with the additional object.